ml-stress-prediction-app/requirements.txt
streamlit==1.28.0
opencv-python==4.8.0
numpy==1.24.3
scikit-image==0.21.0
matplotlib==3.7.2
fpdf2==2.7.0
tensorflow==2.13.0
keras==2.13.0
scikit-learn==1.3.0
pandas==2.0.3
joblib==1.3.1
pillow==10.0.0
pyyaml==6.0